A site project I have been working
Its a website that I have made for a project in design specifically for now that is only consist of HTML and CSS without optimization to different view size. I will soon make it responsive when I learned better with javascript. For now it is just me creating website to practice.